fix: keep agent.middleware out of the webapp import closure
`agent.completion` imported `finalize_agent_run_usage` from the middleware
package, pulling `agent.middleware` into `agent.webapp`'s transitive imports
and failing test_webapp_does_not_import_agent_stack. That closure is pinned
because slow webapp imports delay pod readiness on LangGraph Cloud. Moves the
shared finalizer to `agent_cost`, next to the refresh it schedules.
